Move ZookeeperClient back to kafka.controller
Will do the move+rename in a separate PR
